linuxkit: Update linuxkit to the latest upstream release 1.2.0
With updating linuxkit to a newer release we have to adjust the patches
either way. While at it I reworked them into something that might make
it easier to upstream them at some point.

The update for patch 0001 is inspired by a previous upstreaming attempt
in linuxkit #3446 and the discussion in there.

The update for patch 0002 should make it more clear from the usage that
a non-standard way of using linuxkit is chosen here with introducing a
new output type.

Patch 0003 was carried forward as is to ensure we are not handling the
files that are specified in `replace` twice.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>