kernel: Add a reproducible way to build kernel and nsm module
We want to provide a way to reproducibly build all the binary artifacts
that we provide through aws-nitro-enclaves-cli [0]. This is based on
work previously done by Alexander Graf in [1].

We are using nix to build the kernel image, the nsm module and extract
the final kernel config from the build.

At the same time, we are also bumping the kernel version to the latest
upstream LTS kernel. Instead of building the NSM module as an external
module we pick up the upstream port form kernel series v6.8 through

commit b9873755a6c8 ("misc: Add Nitro Secure Module driver")

[0] https://github.com/aws/aws-nitro-enclaves-cli
[1] https://github.com/agraf/nitro-enclave-kernel
[2] https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=b9873755a6c8

Signed-off-by: Leonard Foerster <foersleo@amazon.com>