linuxkit: Update nixpkgs version used for compilation
Update nixpkgs version to be used for building linuxkit to align between
all subpackages. While we are here also drop exclusion of cgo for
building linuxkit - the exclusion does not seem to have a benefit and
adds extra complexity.

Further rename the nix file in preparation to tie all packages together.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>