github: Add actions flow to test builds and cut releases
We want to automatically test any new PR that gets submitted to ensure
it builds correctly. If someone creates a version tag on the main repo,
we also want to automatically turn that tag into a new release.

This patch adds a github action to

  1) Build x86_64 and aarch64 blobs for each push
  2) Pack up blobs as txz file for each tag push that matches "v*"

With this in place, it should make releases and reasoning about
reproducible builds a lot easier.

Signed-off-by: Alexander Graf <graf@amazon.com>