treewide: Use default compiler and libc from stdenv
The derivations for init, kernel and linuxkit currently consume `gcc`
specifically, and also link against `glibc.static` or musl by passing
explicit linker flags.

There is no particular technical reason for us to enforce this
combination of compiler and libc, as long as the output binaries will
be statically linked for use in an Enclave.

A better pattern is to consume stdenv.cc instead, and leave the choice
of static vs. dynamic linkage up to the nixpkgs package set inwhich the
package is instantiated. The compiler wrapper will pass the appropriate
linker flags for the selected libc, simplifying our build logic.

Using the default compiler and libc also avoids having to build unusual
toolchain combinations when cross-compiling. (The nixpkgs manual [1]
suggests to use the default systems from `pkgs.pkgsCross` which have
pre-built toolchains on cache.nixos.org.)

Replace all direct references to gcc with uses of stdenv.cc, and change
the top level instantiation of the init and linuxkit derivations to
come from `pkgsStatic` instead, which takes care of applying the correct
linker flags to produce statically linked outputs.

[1]: https://nixos.org/manual/nixpkgs/stable/#cross-qa-avoid-compiling-gcc-cross-compiler

Signed-off-by: Simon Veith <sveith@amazon.de>