build: Parametrize pkgs in top-level derivation
The top-level derivation in default.nix currently always uses the
package set pinned in nixpkgs.nix, unlike the sub-derivations which
provide a way of overriding it.

Rename the package set from `nixpkgs` to the commonly used `pkgs` and
turn it into a parameter, which will be useful to instantiate it for
other nixpkgs revisions or configurations (e.g. for cross-compiling).

Signed-off-by: Simon Veith <sveith@amazon.de>