kernel: Add cmdline to the built artifacts (#41)
This commit adds the kernel command line parameters to the built artifacts for the Nitro Enclaves kernel. Specifically, it includes the aarch64.cmdline and x86_64.cmdline files.

As discussed, we may need to further investigate whether these parameters should be kept as-is or modified for each architecture.

The cmdline parameters are coming from the following sources:
- aarch64: https://github.com/aws/aws-nitro-enclaves-cli/blob/main/blobs/aarch64/cmdline
- x86_64: https://github.com/aws/aws-nitro-enclaves-cli/blob/main/blobs/x86_64/cmdline